# Backend Services/Finaco/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Finaco Mutual Fund API Service",
    "description": "These API interfaces will be used by Finaco to interact with customer information.",
    "contact": {
      "email": "kumarvivek@tashiqa.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api-finacomutualfund.azurewebsites.net/api"
    }
  ],
  "paths": {
...
    "/govt/profile/transactions": {
      "get": {
        "tags": [
          "Government"
        ],
        "summary": "Government access to transactions - REQUIRE GOVT BASIC AUTHENTICATION",
        "description": "Get transactions of users.",
        "operationId": "get_govt_transaction",
        "responses": {
          "200": {
            "description": "Succesfull response with transaction."
          },
          "400": {
            "description": "Inavlid body format."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}